+++ requirements-dev.txt
@@ -3,3 +3,4 @@ pytest==6.2.3; python_version > '3'
 pytest-html==1.22.1; python_version <= '2.7'
 pytest-html==3.1.1; python_version > '3'
 pytest-vcr==1.0.2
+pytest-cov==2.12.1
