+++ tests/cassettes/test_wsmtx/test_consultar_comprobante.yaml
@@ -2,12 +2,12 @@ interactions:
 - request:
     body: "<?xml version=\"1.0\" encoding=\"UTF-8\"?><soapenv:Envelope xmlns:soapenv=\"http://schemas.xmlsoap.org/soap/envelope/\"
       xmlns:ser=\"http://impl.service.wsmtxca.afip.gov.ar/service/\">\n<soapenv:Header/>\n<soapenv:Body>\n
-      \   <ser:consultarComprobanteRequest>\n    <authRequest><token>PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iVVRGLTgiIHN0YW5kYWxvbmU9InllcyI/Pgo8c3NvIHZlcnNpb249IjIuMCI+CiAgICA8aWQgc3JjPSJDTj13c2FhaG9tbywgTz1BRklQLCBDPUFSLCBTRVJJQUxOVU1CRVI9Q1VJVCAzMzY5MzQ1MDIzOSIgdW5pcXVlX2lkPSI1MDM4MTE0NyIgZ2VuX3RpbWU9IjE2MjQyMDgzMDMiIGV4cF90aW1lPSIxNjI0MjUxNTYzIi8+CiAgICA8b3BlcmF0aW9uIHR5cGU9ImxvZ2luIiB2YWx1ZT0iZ3JhbnRlZCI+CiAgICAgICAgPGxvZ2luIGVudGl0eT0iMzM2OTM0NTAyMzkiIHNlcnZpY2U9IndzbXR4Y2EiIHVpZD0iU0VSSUFMTlVNQkVSPUNVSVQgMjAyNjc1NjUzOTMsIENOPXJlaW5nYXJ0MjAxOXB1YiIgYXV0aG1ldGhvZD0iY21zIiByZWdtZXRob2Q9IjIyIj4KICAgICAgICAgICAgPHJlbGF0aW9ucz4KICAgICAgICAgICAgICAgIDxyZWxhdGlvbiBrZXk9IjIwMjY3NTY1MzkzIiByZWx0eXBlPSI0Ii8+CiAgICAgICAgICAgIDwvcmVsYXRpb25zPgogICAgICAgIDwvbG9naW4+CiAgICA8L29wZXJhdGlvbj4KPC9zc28+Cg==</token><sign>oe2g4L1Ra3AAsWZcfZrP6rF4BzPa3oYk++zEyRfeq1r/aixKI/WXnczVjPs9/vbl9qlXE3KCK9CvLi6v7UtiMXpGuNdDyntqQnZKyS+bPn/Z9JBvqq7uucYB9R0HBy5AFOC5TVA0rePYkTcPRSCWDMBPtsKaNOO8v8xDCl542tQ=</sign><cuitRepresentada>20267565393</cuitRepresentada></authRequest><consultaComprobanteRequest><codigoTipoComprobante>2</codigoTipoComprobante><numeroPuntoVenta>4000</numeroPuntoVenta><numeroComprobante>286</numeroComprobante></consultaComprobanteRequest></ser:consultarComprobanteRequest>\n</soapenv:Body>\n</soapenv:Envelope>"
+      \   <ser:consultarComprobanteRequest>\n    <authRequest><token>PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iVVRGLTgiIHN0YW5kYWxvbmU9InllcyI/Pgo8c3NvIHZlcnNpb249IjIuMCI+CiAgICA8aWQgc3JjPSJDTj13c2FhaG9tbywgTz1BRklQLCBDPUFSLCBTRVJJQUxOVU1CRVI9Q1VJVCAzMzY5MzQ1MDIzOSIgdW5pcXVlX2lkPSIxNTU2MTM0MTk4IiBnZW5fdGltZT0iMTYyNTE3MDA0OCIgZXhwX3RpbWU9IjE2MjUyMTMzMDgiLz4KICAgIDxvcGVyYXRpb24gdHlwZT0ibG9naW4iIHZhbHVlPSJncmFudGVkIj4KICAgICAgICA8bG9naW4gZW50aXR5PSIzMzY5MzQ1MDIzOSIgc2VydmljZT0id3NtdHhjYSIgdWlkPSJTRVJJQUxOVU1CRVI9Q1VJVCAyMDI2NzU2NTM5MywgQ049cmVpbmdhcnQyMDE5cHViIiBhdXRobWV0aG9kPSJjbXMiIHJlZ21ldGhvZD0iMjIiPgogICAgICAgICAgICA8cmVsYXRpb25zPgogICAgICAgICAgICAgICAgPHJlbGF0aW9uIGtleT0iMjAyNjc1NjUzOTMiIHJlbHR5cGU9IjQiLz4KICAgICAgICAgICAgPC9yZWxhdGlvbnM+CiAgICAgICAgPC9sb2dpbj4KICAgIDwvb3BlcmF0aW9uPgo8L3Nzbz4K</token><sign>Kthm9lNaWyvFqIyKhmN2Hgi29Jyqnx0ZTYe2VD2NHfYmw+Go3A1M9C66VvI2m6DnHYcLoqodfbZwYWY9HxCXtm2sFHrOzARAqkZMCgBcKXFOjRUlK8xGTuO6L2zDN4woF7uQ7XabVEwYXDWZGlTnChLDg5hI96ruzcGLwPjU/QE=</sign><cuitRepresentada>20267565393</cuitRepresentada></authRequest><consultaComprobanteRequest><codigoTipoComprobante>1</codigoTipoComprobante><numeroPuntoVenta>4000</numeroPuntoVenta><numeroComprobante>1835</numeroComprobante></consultaComprobanteRequest></ser:consultarComprobanteRequest>\n</soapenv:Body>\n</soapenv:Envelope>"
     headers:
       accept-encoding:
       - gzip, deflate
       content-length:
-      - '1520'
+      - '1521'
       content-type:
       - text/xml; charset="UTF-8"
       soapaction:
@@ -19,17 +19,19 @@ interactions:
   response:
     body:
       string: <?xml version='1.0' encoding='UTF-8'?><soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/"><soapenv:Body><ns1:consultarComprobanteResponse
-        xmlns:ns1="http://impl.service.wsmtxca.afip.gov.ar/service/"><arrayErrores><codigoDescripcion><codigo>1503</codigo><descripcion>El
-        Comprobante indicado no obra en las bases del Organismo</descripcion></codigoDescripcion></arrayErrores></ns1:consultarComprobanteResponse></soapenv:Body></soapenv:Envelope>
+        xmlns:ns1="http://impl.service.wsmtxca.afip.gov.ar/service/"><comprobante><codigoTipoComprobante>1</codigoTipoComprobante><numeroPuntoVenta>4000</numeroPuntoVenta><numeroComprobante>1835</numeroComprobante><fechaEmision>2021-07-01</fechaEmision><codigoTipoAutorizacion>E</codigoTipoAutorizacion><codigoAutorizacion>71263951788513</codigoAutorizacion><fechaVencimiento>2021-07-11</fechaVencimiento><codigoTipoDocumento>80</codigoTipoDocumento><numeroDocumento>30000000007</numeroDocumento><importeGravado>100.00</importeGravado><importeNoGravado>0.00</importeNoGravado><importeExento>0.00</importeExento><importeSubtotal>100.00</importeSubtotal><importeOtrosTributos>1.00</importeOtrosTributos><importeTotal>122.00</importeTotal><codigoMoneda>PES</codigoMoneda><cotizacionMoneda>1.000000</cotizacionMoneda><observaciones>Observaciones
+        Comerciales, libre</observaciones><codigoConcepto>3</codigoConcepto><fechaServicioDesde>2021-07-01</fechaServicioDesde><fechaServicioHasta>2021-07-01</fechaServicioHasta><fechaVencimientoPago>2021-07-01</fechaVencimientoPago><arrayOtrosTributos><otroTributo><codigo>99</codigo><descripcion>Impuesto
+        Municipal Matanza</descripcion><baseImponible>100.00</baseImponible><importe>1.00</importe></otroTributo></arrayOtrosTributos><arrayItems><item><unidadesMtx>123456</unidadesMtx><codigoMtx>1234567890123</codigoMtx><codigo>P0001</codigo><descripcion>Descripcion
+        del producto P0001</descripcion><cantidad>2.000000</cantidad><codigoUnidadMedida>7</codigoUnidadMedida><precioUnitario>100.000000</precioUnitario><importeBonificacion>0.000000</importeBonificacion><codigoCondicionIVA>5</codigoCondicionIVA><importeIVA>42.00</importeIVA><importeItem>242.00</importeItem></item><item><descripcion>bonificacion</descripcion><codigoUnidadMedida>99</codigoUnidadMedida><codigoCondicionIVA>5</codigoCondicionIVA><importeIVA>-21.00</importeIVA><importeItem>-121.00</importeItem></item></arrayItems><arraySubtotalesIVA><subtotalIVA><codigo>5</codigo><importe>21.00</importe></subtotalIVA></arraySubtotalesIVA></comprobante></ns1:consultarComprobanteResponse></soapenv:Body></soapenv:Envelope>
     headers:
       Connection:
       - close
       Content-Type:
       - text/xml;charset=UTF-8
       Date:
-      - Sun, 20 Jun 2021 17:00:04 GMT
+      - Thu, 01 Jul 2021 20:15:46 GMT
       Set-Cookie:
-      - TS01761d9e=01439f1ddff2ff9a2dcbd9996685001529f2988ecfa7dfceae3b8c18f760ac2f079e3af5a1;
+      - TS01761d9e=01439f1ddf4ba5e474c80782372cb77d7660f6d4ab67797dede81bb3bea682669414d4559d;
         Path=/
       Transfer-Encoding:
       - chunked
